fix(models): satisfy CLI maintainability gates
